+++ Sources/XcodeGenKit/CarthageDependencyResolver.swift
@@ -39,7 +39,7 @@ public class CarthageDependencyResolver {
         switch linkType {
         case .static:
             return "\(buildPath)/\(platform.carthageName)/Static"
-        case .dynamic:
+        case .dynamic, .staticBinary:
             return "\(buildPath)/\(platform.carthageName)"
         }
     }